+++ mtools/mgenerate/operators/operators.py
@@ -8,6 +8,7 @@ from dateutil import parser
 
 import time
 import string
+import itertools
 
 
 class BaseOperator(object):
@@ -18,7 +19,6 @@ class BaseOperator(object):
 
     def __init__(self, decode_method):
         self._decode = decode_method
-
 
     def _parse_options(self, options={}):
         parsed = self.defaults.copy()
@@ -84,19 +84,25 @@ class FloatOperator(BaseOperator):
         return val
 
 
-
 class IncOperator(BaseOperator):
 
-    dict_format = False
+    dict_format = True
     string_format = True
     names = ['$inc']
-    value = -1
+    defaults = OrderedDict([ ('start', 0), ('step', 1) ])
+    
+    def __init__(self, decode_method):
+        self.counter = None
+        BaseOperator.__init__(self, decode_method)
 
     def __call__(self, options=None):
         options = self._parse_options(options)
+        
+        # initialize counter on first use (not threadsafe!)
+        if self.counter == None:
+            self.counter = itertools.count(options['start'], options['step'])
 
-        self.value += 1
-        return self.value
+        return self.counter.next()
 
 
 class StringOperator(BaseOperator):
@@ -191,6 +197,42 @@ class ArrayOperator(BaseOperator):
 
         # build array of 'of' elements, but don't evaluate them yet
         return [ options['of'] ] * number
+
+
+class CoordinateOperator(BaseOperator):
+
+    dict_format = True
+    string_format = True
+    names = ['$coordinates', '$coordinate', '$coord']
+    defaults = OrderedDict([ ('long_lim', [-180, 180]), ('lat_lim', [-90, 90]) ])
+
+    def __call__(self, options=None):
+        options = self._parse_options(options)
+
+        # evaluate limits
+        long_lim = self._decode(options['long_lim'])
+        lat_lim = self._decode(options['lat_lim'])
+
+        # return coordinate by using random numbers between limits
+        return [ {"$float": long_lim }, {"$float": lat_lim }]
+
+
+class PointOperator(BaseOperator):
+
+    dict_format = True
+    string_format = True
+    names = ['$point']
+    defaults = OrderedDict([ ('long_lim', [-180, 180]), ('lat_lim', [-90, 90]) ])
+
+    def __call__(self, options=None):
+        options = self._parse_options(options)
+
+        # evaluate limits
+        long_lim = self._decode(options['long_lim'])
+        lat_lim = self._decode(options['lat_lim'])
+
+        # return coordinate by using random numbers between limits
+        return { "type": "Point", "coordinates": { "$coord": [long_lim, lat_lim] } }
 
 
 class DateTimeOperator(BaseOperator):